@@ -45,11 +45,12 @@
     "react-dropzone": "^10.2.1"
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.0.0",
-    "@material-ui/icons": "^4.0.0",
+    "@mui/icons-material": "^5.0.0",
+    "@mui/material": "^5.0.0",
+    "@mui/styles": "^5.0.1",
     "prop-types": "^15.7.2",
-    "react": ">= 16.8",
-    "react-dom": ">= 16.8"
+    "react": ">= 17.0",
+    "react-dom": ">= 17.0"
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
@@ -60,8 +61,8 @@
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.10.1",
-    "@material-ui/core": "^4.9.7",
-    "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "^5.0.1",
+    "@mui/material": "^5.0.2",
     "@rollup/plugin-commonjs": "^11.0.2",
     "@rollup/plugin-node-resolve": "^7.1.1",
     "babel-eslint": "^10.1.0",
@@ -73,8 +74,8 @@
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-react": "^7.19.0",
     "gh-pages": "^1.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-styleguidist": "^11.0.4",
     "rollup": "^1.32.0",
     "rollup-plugin-babel": "^4.4.0",
