@@ -1,8 +1,9 @@
-import Snackbar from '@material-ui/core/Snackbar';
-import Typography from '@material-ui/core/Typography';
-import {withStyles} from '@material-ui/core/styles';
-import AttachFileIcon from '@material-ui/icons/AttachFile';
-import CloudUploadIcon from '@material-ui/icons/CloudUpload';
+import Snackbar from '@mui/material/Snackbar';
+import Typography from '@mui/material/Typography';
+import Button from '@material-ui/core/Button';
+import {withStyles} from '@mui/styles';
+import AttachFileIcon from '@mui/icons-material/AttachFile';
+import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import * as React from 'react';
@@ -57,6 +58,10 @@ const styles = ({palette, shape, spacing}) => ({
         width: 51,
         height: 51,
         color: palette.text.primary,
+    },
+    resetButton: {
+        display: 'block',
+        margin: '10px 0',
     },
 });
 
@@ -221,6 +226,7 @@ class DropzoneAreaBase extends React.PureComponent {
             showPreviews,
             showPreviewsInDropzone,
             useChipsForPreview,
+            reset,
         } = this.props;
         const {openSnackBar, snackbarMessage, snackbarVariant} = this.state;
 
@@ -290,6 +296,20 @@ class DropzoneAreaBase extends React.PureComponent {
                     )}
                 </Dropzone>
 
+                {
+                    reset && (
+                        React.isValidElement(reset) ?
+                            reset :
+                            <Button
+                                onClick={reset.onClick}
+                                variant="outlined"
+                                className={classes.resetButton}
+                            >
+                                {reset.text || 'reset'}
+                            </Button>
+                    )
+                }
+
                 {previewsVisible &&
                     <Fragment>
                         <Typography variant="subtitle1" component="span">
@@ -347,6 +367,7 @@ DropzoneAreaBase.defaultProps = {
     previewChipProps: {},
     previewGridClasses: {},
     previewGridProps: {},
+    reset: undefined,
     showAlerts: true,
     alertSnackbarProps: {
         anchorOrigin: {
@@ -431,6 +452,21 @@ DropzoneAreaBase.propTypes = {
     previewGridProps: PropTypes.object,
     /** The label for the file preview section. */
     previewText: PropTypes.string,
+    /**
+     * The node of button to clear dropzone.
+     *
+     * - can be a node to mount in a placeholder.
+     * - can be an object:
+     *  - text (string) - text of the button
+     *  - onClick (function) - callback fired when reset button clicked
+     */
+    reset: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.shape({
+            text: PropTypes.string,
+            onClick: PropTypes.func,
+        }),
+    ]),
     /**
      * Shows styled Material-UI Snackbar when files are dropped, deleted or rejected.
      *
